chore: move DensityEstimatorBuilder to estimators
