+++ sbi/inference/trainers/nle/mnle.py
@@ -12,9 +12,9 @@ from sbi.inference.posteriors.posterior_parameters import (
     RejectionPosteriorParameters,
     VIPosteriorParameters,
 )
-from sbi.inference.trainers.base import DensityEstimatorBuilder
 from sbi.inference.trainers.nle.nle_base import LikelihoodEstimatorTrainer
 from sbi.neural_nets.estimators import MixedDensityEstimator
+from sbi.neural_nets.estimators.base import DensityEstimatorBuilder
 from sbi.sbi_types import TensorboardSummaryWriter
 from sbi.utils.sbiutils import del_entries
