+++ sbi/inference/trainers/nle/nle_base.py
@@ -23,9 +23,10 @@ from sbi.inference.posteriors.posterior_parameters import (
 )
 from sbi.inference.potentials import likelihood_estimator_based_potential
 from sbi.inference.potentials.likelihood_based_potential import LikelihoodBasedPotential
-from sbi.inference.trainers.base import DensityEstimatorBuilder, NeuralInference
+from sbi.inference.trainers.base import NeuralInference
 from sbi.neural_nets import likelihood_nn
 from sbi.neural_nets.estimators import ConditionalDensityEstimator
+from sbi.neural_nets.estimators.base import DensityEstimatorBuilder
 from sbi.neural_nets.estimators.shape_handling import (
     reshape_to_batch_event,
 )
